\name{genomicsdb-package}
\alias{genomicsdb-package}
\alias{genomicsdb}
\docType{package}
\title{
  GenomicsDB R Interface
}
\description{
  R Interface via bindings to the native GenomicsDB library
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
GenomicsDB OpenSource
Maintainer:  Support<support@genomicsdb.org>
}
\references{
        github:GenomicsDB/GenomicsDB
}
\examples{
  \dontrun{

     initialize()
     query()
  }
}
